minor various changes

+++ scripts/nerdfont_setup.sh
@@ -10,14 +10,18 @@ if ! [ -d ~/.local/share/fonts ]; then
 fi
 
 # Download and install the font
-mkdir tmp/
-curl -OL https://github.com/ryanoasis/nerd-fonts/releases/latest/download/${FONTNAME}.tar.xz --output-dir tmp/
-tar -xf ${FONTNAME}.tar.xz -C tmp/
-mv tmp/*.ttf ~/.local/share/fonts
-rm -rf tmp/
 
-# Update the font cache
-sudo fc-cache -fv
+if ! [ -f ~/.local/share/fonts/${FONTNAME}* ]; then
+	mkdir tmp/
+	curl -OL https://github.com/ryanoasis/nerd-fonts/releases/latest/download/${FONTNAME}.tar.xz --output-dir tmp/
+	tar -xf ${FONTNAME}.tar.xz -C tmp/
+	mv tmp/*.ttf ~/.local/share/fonts
+	rm -rf tmp/
 
-echo "${FONTNAME} nerdfont installed successfully"
+	# Update the font cache
+	sudo fc-cache -fv
 
+	echo "${FONTNAME} installed successfully"
+else
+	echo "${FONTNAME} already installed"
+fi
